chore: restructure protocols
